Using floats (double) for more of the edge calculations...
More precision is probably better?

+++ src/lib/extractor/Midpoints.h
@@ -6,8 +6,8 @@
 class Midpoints
 {
 public:
-	Midpoints(const std::vector<point<int>>& points)
-		: _points(points)
+	Midpoints(const std::vector<point<double>>& points)
+	    : _points(points)
 	{
 	}
 
@@ -16,26 +16,26 @@ public:
 		return _points.size() < 4;
 	}
 
-	const point<int>& top() const
+	const point<double>& top() const
 	{
 		return _points[0];
 	}
 
-	const point<int>& bottom() const
+	const point<double>& bottom() const
 	{
 		return _points[1];
 	}
 
-	const point<int>& left() const
+	const point<double>& left() const
 	{
 		return _points[2];
 	}
 
-	const point<int>& right() const
+	const point<double>& right() const
 	{
 		return _points[3];
 	}
 
 protected:
-	std::vector<point<int>> _points;
+	std::vector<point<double>> _points;
 };
